# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: mosorio@isi.edu
  description: This is MINT Model Catalog You can find out more about     Model Catalog at [https://mintproject.github.io/Mint-ModelCatalog-Ontology/modelCatalog/release/0.4.0/](https://mintproject.github.io/Mint-ModelCatalog-Ontology/modelCatalog/release/0.4.0/)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Model Catalog
  version: 0.0.2
...
servers:
- url: https://api.models.mint.isi.edu/v0.0.2
- url: https://dev.api.models.mint.isi.edu/v0.0.2
- url: http://localhost:8080/v0.0.2
paths:
...
  /modelconfigurations:
    get:
      operationId: list_model_configurations
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelConfiguration'
                type: array
          description: successful operation
        400:
          description: Invalid uri supplied
      security: []
      summary: List modelconfiguration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    post:
      operationId: add_model_configuration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelConfiguration'
        required: true
      responses:
        201:
          content:
            application/json: {}
          description: Created
        400:
          description: Invalid
      security:
      - BearerAuth: []
      summary: Create a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
  /modelconfiguration/{id}/inputs:
    get:
      operationId: get_inputs_by_modelconfiguration
      parameters:
      - description: |
          The name of the resource
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiResponse'
                type: array
          description: successful operation
        400:
          description: Invalid uri supplied
        404:
          description: Model Configuration not found
      security: []
      summary: Get the inputs of a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    post:
      operationId: add_inputs_by_modelconfiguration
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/DataSet'
              type: array
        required: true
      responses:
        201:
          description: Created
        400:
          description: Request body is incorrect
        403:
          content:
            application/json:
              examples:
                exit:
                  value: |-
                    {
                        "code": 30,
                        "type": "some text",
                        "message": "some text"
                    }
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: You are not authorized to do that
      security:
      - BearerAuth: []
      summary: Creates a new instance of a `Dataset` related as Input.
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
  /modelconfiguration/{id}/outputs:
    get:
      operationId: get_outputs_by_modelconfiguration
      parameters:
      - description: |
          The name of the resource
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiResponse'
                type: array
          description: successful operation
        400:
          description: Invalid uri supplied
        404:
          description: Model Configuration not found
      security: []
      summary: Get the outputs of a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    post:
      operationId: modelconfiguration_id_outputs_post
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/DataSet'
              type: array
        required: true
      responses:
        201:
          description: Created
        400:
          description: Request body is incorrect
        403:
          content:
            application/json:
              examples:
                exit:
                  value: |-
                    {
                        "code": 30,
                        "type": "some text",
                        "message": "some text"
                    }
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: You are not authorized to do that
      security:
      - BearerAuth: []
      summary: Create the output of a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
  /modelconfiguration/{id}/parameters:
    get:
      operationId: get_parameters_by_modelconfiguration
      parameters:
      - description: |
          The name of the resource
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiResponse'
                type: array
          description: successful operation
        400:
          description: Invalid uri supplied
        404:
          description: Model Configuration not found
      security: []
      summary: Get the parameters of a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    post:
      description: Creates a new instance of a `Dataset` and it related with the `ModelConfiguration`.
      operationId: add_parameters_by_modelconfiguration
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Parameter'
              type: array
        required: true
      responses:
        201:
          description: Created
        400:
          description: Request body is incorrect
        403:
          content:
            application/json:
              examples:
                exit:
                  value: |-
                    {
                        "code": 30,
                        "type": "some text",
                        "message": "some text"
                    }
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: You are not authorized to do that
      security:
      - BearerAuth: []
      summary: Create the inputs of a model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
...
  /models:
    get:
      description: Gets a list of all `model` entities.
      operationId: getmodels
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Model'
                type: array
          description: Successful response - returns an array of `model` entities.
      security: []
      summary: List All models
      tags:
      - model
      x-openapi-router-controller: openapi_server.controllers.model_controller
    post:
      description: Creates a new instance of a `model`.
      operationId: createmodel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Model'
        description: A new `model` to be created.
        required: true
      responses:
        201:
          description: Successful response.
      security:
      - BearerAuth: []
      summary: Create a model
      tags:
      - model
      x-openapi-router-controller: openapi_server.controllers.model_controller
...
  /model/{modelId}:
    delete:
      description: Deletes an existing `Model`.
      operationId: delete_model
      parameters:
      - description: A unique identifier for a `Model`.
        explode: false
        in: path
        name: modelId
        required: true
        schema:
          type: string
        style: simple
      responses:
        204:
          description: Successful response.
      security:
      - BearerAuth: []
      summary: Delete a Model
      tags:
      - model
      x-openapi-router-controller: openapi_server.controllers.model_controller
    get:
      description: Gets the details of a single instance of a `Model`.
      operationId: get_model
      parameters:
      - description: A unique identifier for a `Model`.
        explode: false
        in: path
        name: modelId
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
          description: Successful response - returns a single `Model`.
      security: []
      summary: Get a Model
      tags:
      - model
      x-openapi-router-controller: openapi_server.controllers.model_controller
    put:
      description: Updates an existing `Model`.
      operationId: update_model
      parameters:
      - description: A unique identifier for a `Model`.
        explode: false
        in: path
        name: modelId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Model'
        description: Updated `Model` information.
        required: true
      responses:
        202:
          description: Successful response.
      security:
      - BearerAuth: []
      summary: Update a Model
      tags:
      - model
      x-openapi-router-controller: openapi_server.controllers.model_controller
...
  /modelconfiguration/{id}:
    delete:
      description: Deletes an existing `ModelConfiguration`.
      operationId: delete_model_configuration
      parameters:
      - description: A unique identifier for a `ModelConfiguration`.
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        204:
          description: Successful response.
        400:
          description: Invalid ID supplied
        403:
          description: You don't the permissions
      security:
      - BearerAuth: []
      summary: Delete a ModelConfiguration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    get:
      description: Gets the details of a single instance of a `ModelConfiguration`.
      operationId: get_model_configuraton_by_uri
      parameters:
      - description: A unique identifier for a `ModelConfiguration`.
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelConfiguration'
          description: successful operation
        400:
          description: Invalid uri supplied
        404:
          description: Model Configuration not found
      security: []
      summary: Get modelconfiguration by uri
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
    put:
      operationId: update_model_configuration
      parameters:
      - description: A unique identifier for a `ModelConfiguration`.
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: A unique identifier for a `ModelConfiguration`.
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelConfiguration'
        required: true
      responses:
        200:
          description: Updated
      security:
      - BearerAuth: []
      summary: Update model configuration
      tags:
      - modelconfiguration
      x-openapi-router-controller: openapi_server.controllers.modelconfiguration_controller
components:
  schemas:
    DataSet:
      example:
        id: crop-yield
        label: crop-yield
        has_format: csv
        description: Contains the yield elasticities to fertilizer use that are derived from Cycles output
        has_dimensionality: 0
        has_presentation:
        - id: https://w3id.org/mint/instance/econ_ybarN
      properties:
        label:
          type: string
        has_format:
          type: string
        description:
          type: string
        has_dimensionality:
          format: int64
          type: integer
        has_presentation:
          items:
            $ref: '#/components/schemas/VariablePresentation'
          type: array
        type:
          items:
            type: string
          type: array
        id:
          type: string
      required:
      - id
      type: object
    ModelConfiguration:
      example:
        id: economic
        label: Economic aggregate crop supply
        has_input:
        - id: econ_yield
        has_output:
        - id: econ_land_use
        description: Aggregate crop supply response model for the country of South Sudan configuration
        has_cag:
        - id: economic_aggregate_crop_supply_CAG_variables
        - id: economic_aggregate_crop_supply_CAG_process
        has_process:
        - id: crop_yield
        - id: crop_supply
        has_time_interval:
        - id: economic_aggregate_crop_supply_TI
        has_implementation_script_location: https://raw.githubusercontent.com/KnowledgeCaptureAndDiscovery/MINT-WorkflowDomain/master/WINGSWorkflowComponents/economic/run
        has_container: https://w3id.org/mint/instance/mintproject/economic:latest
        has_constraint: '[ backwardPIHMInputsDataHaveSameRegion: (?c rdf:type acdom:pihmClass)
          (?c ac:hasInput ?in1) (?c ac:hasInput ?in2) noValue(?in1 dcdom:region) noValue(?in2
          dcdom:region) uriConcat(?c ''_region'' ?location) -> (?in1 dcdom:region
          ?location) (?in2 dcdom:region ?location) print(''Setting domain of ?in1
          and ?in2 to ?location because both inputs to PIHM should be from the same
          region'')]'
        has_parameter:
        - id: pihm_start-date
        - id: pihm_end-date
        has_component_location: https://github.com/KnowledgeCaptureAndDiscovery/MINT-WorkflowDomain/raw/master/WINGSWorkflowComponents/economic/economic.zip
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        label:
          type: string
        has_input:
          items:
            $ref: '#/components/schemas/DataSet'
          type: array
        has_output:
          items:
            $ref: '#/components/schemas/DataSet'
          type: array
        description:
          type: string
        has_cag:
          items:
            $ref: '#/components/schemas/CAG'
          type: array
        has_process:
          items:
            $ref: '#/components/schemas/Process'
          type: array
        has_time_interval:
          items:
            $ref: '#/components/schemas/TimeInterval'
          type: array
        has_implementation_script_location:
          type: string
        has_container:
          type: string
        has_constraint:
          type: string
        has_parameter:
          items:
            $ref: '#/components/schemas/Parameter'
          type: array
        has_component_location:
          type: string
      required:
      - id
      type: object
    Parameter:
      example:
        id: pihm_start-date
        type:
        - https://w3id.org/wings/export/MINT#StartDate
        label: pihm_start-date
        description: start date
        has_default_value: 2017-01-01
        has_data_type: string
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        label:
          type: string
        description:
          type: string
        has_default_value:
          type: string
        has_data_type:
          type: string
      required:
      - id
      type: object
    Process:
      example:
        id: crop_yield
        label: crop yield
        influences:
        - id: https://w3id.org/mint/instance/crop_supply
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        label:
          type: string
        influences:
          description: Property that describe the causal relationship between two processes, two variables, a process and a variable or a variable and a process
          items:
            $ref: '#/components/schemas/Process'
          type: array
      required:
      - id
      type: object
    TimeInterval:
      example:
        id: economic_aggregate_crop_supply_TI
        type:
        - https://example.org/externalCatalog#TimeInterval
        label: Simple economic time interval
        description: Time interval used in the aggregate crop supply response model for the country of South Sudan configuration
        has_unit:
          id: https://w3id.org/mint/instance/yearT
        has_value: 1
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        label:
          type: string
        description:
          type: string
        has_unit:
          $ref: '#/components/schemas/Unit'
        has_value:
          format: int64
          type: integer
      required:
      - id
      type: object
    Unit:
      example:
        id: USD_kg_1M_1
        type:
        - http://qudt.org/1.1/schema/qudt#Unit
        label: USD kg-1
        ccut_has_part:
        - u_USD
        - u__1_M_kilo_gram_kg
        ccut_has_dimensionality:
        - M-1
        abbreviation:
        - USD kg-1
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        description:
          type: string
        label:
          type: string
        ccut_has_part:
          items:
            type: string
          type: array
        ccut_has_dimensionality:
          items:
            type: string
          type: array
        abbreviation:
          items:
            type: string
          type: array
      required:
      - id
      type: object
    VariablePresentation:
      example:
        id: econ_ybarN
        label: ybarN
        description: Contains the yield elasticities to fertilizer use that are derived from Cycles output
        has_long_name: yield elasticity
        has_short_name: ybarN
        has_standard_variable: crop_fertilizer~nitrogen__yield_elasticity
        has_relevance_level: 1
        uses_unit:
          id: USD_kg_1M_1
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        label:
          type: string
        description:
          type: string
        has_long_name:
          type: string
        has_short_name:
          type: string
        has_standard_variable:
          type: string
        has_relevance_level:
          format: int64
          type: integer
        uses_unit:
          $ref: '#/components/schemas/Unit'
      required:
      - id
      type: object
...
    ApiResponse:
      example:
        code: 0
        type: type
        message: message
      properties:
        code:
          format: int32
          type: integer
        type:
          type: string
        message:
          type: string
      type: object
    CAG:
      example:
        id: economic_aggregate_crop_supply_CAG_variables
        label: CAG for Economic model for aggregate crop supply response model variables for the country of South Sudan configuration
        type:
        - https://w3id.org/mint/modelCatalog#CAG
        has_part:
        - id: https://w3id.org/mint/instance/econ_ybarN
        - id: https://w3id.org/mint/instance/econ_production
      properties:
        id:
          type: string
        type:
          items:
            type: string
          type: array
        has_part:
          description: Array of Parameter or Variables
          items:
            $ref: '#/components/schemas/Parameter'
          type: array
        label:
          type: string
      required:
      - id
      type: object
    Model:
      example:
        id: DSSAT
        label: DSSAT
        description: The Decision Support System for Agrotechnology Transfer (DSSAT) comprises dynamic crop growth simulation model for over 40 crops. The model simulates growth; development; and yield as a function of the soil-plant-atmosphere dynamics.
        type:
        - https://w3id.org/mint/modelCatalog#Model
        - https://w3id.org/mint/modelCatalog#TheoryBasedModel
        has_software_version:
        - DSSAT_4.7
        has_model_category:
        - Agriculture
        has_documentation:
        - https://dssat.net
      properties:
        id:
          type: string
        label:
          type: string
        type:
          items:
            type: string
          type: array
        has_software_version:
          items:
            type: string
          type: array
        has_model_category:
          items:
            type: string
          type: array
        has_documentation:
          items:
            type: string
          type: array
      required:
      - id
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.user_controller.decode_token